Move S6797 into S6803 implementation
