+++ analyzers/tests/SonarAnalyzer.UnitTest/TestCases/BlazorQueryParameterTypeShouldBeSupported.NoRoute.razor
@@ -1,79 +1,53 @@
-﻿<p>
-    Component parameters supplied from the query string support the following types:
-    <ul>
-        <li>bool, DateTime, decimal, double, float, Guid, int, long, string.</li>
-        <li>Nullable variants of the preceding types.</li>
-        <li>Arrays of the preceding types, whether they're nullable or not nullable.</li>
-    </ul>
-</p>
-
-@code {
+﻿@code {
     [Parameter]
-    [SupplyParameterFromQuery]
     public bool BoolParam { get; set; } // Compliant
 
     [Parameter]
-    [SupplyParameterFromQuery]
     public DateTime DateTimeParam { get; set; } // Compliant
 
     [Parameter]
-    [SupplyParameterFromQuery]
     public decimal DecimalParam { get; set; } // Compliant
 
     [Parameter]
-    [SupplyParameterFromQuery]
     public double DoubleParam { get; set; } // Compliant
 
     [Parameter]
-    [SupplyParameterFromQuery]
     public float FloatParam { get; set; } // Compliant
 
     [Parameter]
-    [SupplyParameterFromQuery]
     public Guid GuidParam { get; set; } // Compliant
 
     [Parameter]
-    [SupplyParameterFromQuery]
     public int IntParam { get; set; } // Compliant
 
     [Parameter]
-    [SupplyParameterFromQuery]
     public long LongParam { get; set; } // Compliant
 
     [Parameter]
-    [SupplyParameterFromQuery]
     public string StringParam { get; set; } // Compliant
 
     [Parameter]
-    [SupplyParameterFromQuery]
     public bool? NullableBoolParamShortForm { get; set; } // Compliant
 
     [Parameter]
-    [SupplyParameterFromQuery]
     public Nullable<bool> NullableBoolParamLongForm { get; set; } // Compliant
 
     [Parameter]
-    [SupplyParameterFromQuery]
     public bool?[] ArrayOfNullableBoolParamShortForm { get; set; } // Compliant
 
     [Parameter]
-    [SupplyParameterFromQuery]
     public Nullable<bool>[] ArrayOfNullableBoolParamLongForm { get; set; } // Compliant
 
     [Parameter]
-    [SupplyParameterFromQuery]
     public TimeSpan TimeSpan { get; set; } // Compliant
 
     [Parameter]
-    [SupplyParameterFromQuery]
     public TimeSpan? NullableTimeSpan { get; set; } // Compliant
 
     [Parameter]
-    [SupplyParameterFromQuery]
     public Nullable<TimeSpan> NullableTimeSpanLongForm { get; set; } // Compliant
 
     [Parameter]
-    [SupplyParameterFromQuery]
     public TimeSpan[] TimeSpanArrayLongForm { get; set; } // Compliant
 
     [Parameter]
