+++ analyzers/tests/SonarAnalyzer.UnitTest/TestCases/SupplyParameterFromQueryNeedRoutableComponent_Partial.cs
@@ -1,0 +1,12 @@
+﻿using Microsoft.AspNetCore.Components;
+using System;
+
+namespace EmptyProject
+{
+    public partial class SupplyParameterFromQueryNeedRoutableComponent_Partial : ComponentBase
+    {
+        [Parameter]
+        [SupplyParameterFromQuery]
+        public int Number { get; set; } // Compliant
+    }
+}
